requirements.txt
flask
requests
websocket-client
psutil